feat(mcp): extract platform helpers into platforms-lib
Move MCP platform feed slug and skill compatibility helpers into a
focused lib module with a thin re-export wrapper and targeted tests.
